refactor(liqonet utils): move netlink dependency out of liqonet utils package
This is needed because liqoctl connect command imports the ipam package
which depends on the utils package. This causet the build of liqoctl
command for windows to fail.
