+++ cmd/liqonet/gateway-operator.go
@@ -31,6 +31,7 @@ import (
 	liqoconst "github.com/liqotech/liqo/pkg/consts"
 	liqonetns "github.com/liqotech/liqo/pkg/liqonet/netns"
 	"github.com/liqotech/liqo/pkg/liqonet/utils"
+	"github.com/liqotech/liqo/pkg/liqonet/utils/links"
 	"github.com/liqotech/liqo/pkg/utils/mapper"
 	"github.com/liqotech/liqo/pkg/utils/restcfg"
 )
@@ -144,7 +145,7 @@ func runGatewayOperator(commonFlags *liqonetCommonFlags, gatewayFlags *gatewayOp
 			klog.Errorf("an error occurred while deleting netns {%s}: %v", liqoconst.GatewayNetnsName, err)
 		}
 		klog.Info("cleaning up wireguard tunnel interface")
-		if err := utils.DeleteIFaceByName(liqoconst.DeviceName); err != nil {
+		if err := links.DeleteIFaceByName(liqoconst.DeviceName); err != nil {
 			klog.Errorf("an error occurred while deleting iface {%s}: %v", liqoconst.DriverName, err)
 		}
 		os.Exit(1)
