+++ pkg/liqonet/utils/links/links.go
@@ -1,0 +1,43 @@
+// Copyright 2019-2022 The Liqo Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package links
+
+import (
+	"errors"
+
+	"github.com/vishvananda/netlink"
+	"k8s.io/klog/v2"
+)
+
+// DeleteIFaceByName deletes the interface that has the given name.
+func DeleteIFaceByName(ifaceName string) error {
+	existingIface, err := netlink.LinkByName(ifaceName)
+	if err != nil {
+		if errors.As(err, &netlink.LinkNotFoundError{}) {
+			return nil
+		}
+		klog.Errorf("an error occurred while getting network interface {%s}: %v", ifaceName, err)
+		return err
+	}
+	// Remove the existing network interface.
+	if err = netlink.LinkDel(existingIface); err != nil && errors.As(err, &netlink.LinkNotFoundError{}) {
+		if errors.As(err, &netlink.LinkNotFoundError{}) {
+			return nil
+		}
+		klog.Errorf("an error occurred while deleting network interface {%s}: %v", existingIface.Attrs().Name, err)
+		return err
+	}
+	return nil
+}
