+++ pkg/liqonet/utils/utils_test.go
@@ -20,7 +20,6 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
-	"github.com/vishvananda/netlink"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
@@ -34,7 +33,6 @@ const (
 	labelValue        = "LabelValue"
 	annotationKey     = "net.liqo.io/AnnotationKey"
 	annotationValue   = "AnnotationValue"
-	interfaceName     = "dummy-link"
 )
 
 var (
@@ -235,29 +233,4 @@ var _ = Describe("Liqonet", func() {
 			})
 		})
 	})
-
-	Describe("testing DeleteIfaceByName function", func() {
-		Context("when network interface exists", func() {
-			BeforeEach(func() {
-				// Create dummy link.
-				err := netlink.LinkAdd(&netlink.Dummy{LinkAttrs: netlink.LinkAttrs{Name: interfaceName}})
-				Expect(err).ShouldNot(HaveOccurred())
-			})
-
-			It("should return nil", func() {
-				err := utils.DeleteIFaceByName(interfaceName)
-				Expect(err).Should(BeNil())
-				_, err = netlink.LinkByName(interfaceName)
-				Expect(err).Should(MatchError("Link not found"))
-			})
-		})
-
-		Context("when network interface does not exist", func() {
-			It("should return nil", func() {
-				err := utils.DeleteIFaceByName("not-existing")
-				Expect(err).Should(BeNil())
-			})
-		})
-	})
-
 })
